fix: change esm.mjs
